xxxx/gui_competition_eval/code-for-student/requirements.txt
openai==2.24.0
pillow==12.1.1
